backend update

+++ banking_system_backend/api/views.py
@@ -1,59 +1,74 @@
 import json
-
 from django.http import JsonResponse
-from django.views.decorators.csrf import csrf_exempt
-from django.views.generic import View
 
 from api.models import User, Token
 
 
-class login(View):
-    @csrf_exempt
-    def post(self, request):
-        print(request.body)
+def login(request):
+    # 检查请求方法是否正确
+    if request.method == 'POST':
+        # 获取用户名与密码
         data = json.loads(request.body)
         username = data['username']
         password = data['password']
+        # 检查用户是否存在
         try:
             user = User.objects.get(username=username)
         except User.DoesNotExist:
             return JsonResponse({'message': 'Invalid login credentials', 'status': 'error'})
+        # 检查密码是否正确
         if user.check_password(password):
             token = Token.objects.get_or_create(user=user)
             return JsonResponse({'message': 'Login successful', 'status': 'success', 'accountId': user.id,
                                  'sessionId': str(token[0]), 'userType': user.userType})
         else:
             return JsonResponse({'message': 'Invalid login credentials', 'status': 'error'})
+    else:
+        return JsonResponse({'message': 'Invalid request method', 'status': 'error'})
 
 
-class logout(View):
-    @csrf_exempt
-    def post(self, request):
+def logout(self, request):
+    if request.method == 'POST':
         # 获取token
-        auth_token = request.META.get('HTTP_AUTHORIZATION').split(' ')[1]
+        temp = request.META.get('HTTP_AUTHORIZATION').split(' ')
+        if len(temp) > 1:
+            auth_token = request.META.get('HTTP_AUTHORIZATION').split(' ')[1]
+        else:
+            return JsonResponse({'message': 'Invalid AUTHORIZATION', 'status': 'error'})
+        # 删除token
         try:
             token = Token.objects.get(key=auth_token)
             token.delete()
             return JsonResponse({'message': 'Logout successful', 'status': 'success'})
         except Token.DoesNotExist:
             return JsonResponse({'message': 'Invalid AUTHORIZATION', 'status': 'error'})
+    else:
+        return JsonResponse({'message': 'Invalid request method', 'status': 'error'})
 
 
-class createAccount(View):
-    @csrf_exempt
-    def post(self, request):
+def create_account(self, request):
+    # 检查请求方法是否正确
+    if request.method == 'POST':
         # 获取token
-        auth_token = request.META.get('HTTP_AUTHORIZATION').split(' ')[1]
+        temp = request.META.get('HTTP_AUTHORIZATION').split(' ')
+        if len(temp) > 1:
+            auth_token = request.META.get('HTTP_AUTHORIZATION').split(' ')[1]
+        else:
+            return JsonResponse({'message': 'Invalid AUTHORIZATION', 'status': 'error'})
+        # 寻找token对应的用户
         try:
             token = Token.objects.get(key=auth_token)
             current_user = token.user
         except Token.DoesNotExist:
             return JsonResponse({'message': 'Invalid AUTHORIZATION', 'status': 'error'})
+        # 检查当前用户是否为clerk
         if current_user.is_clerk():
+            # 获取新用户的信息
             data = json.loads(request.body)
             username = data['username']
             password = data['password']
             userType = data['userType']
+            # 检查用户是否已经存在
             try:
                 user = User.objects.get(username=username)
             except User.DoesNotExist:
@@ -62,19 +77,27 @@ class createAccount(View):
                                      'status': 'success', 'accountId': user.id})
             return JsonResponse({'message': 'Username already exists', 'status': 'error'})
         else:
-            return JsonResponse({'message': 'Current User is not clerk or admin', 'status': 'error'})
+            return JsonResponse({'message': 'Current User is not clerk', 'status': 'error'})
+    else:
+        return JsonResponse({'message': 'Invalid request method', 'status': 'error'})
 
 
-@csrf_exempt
 def delete_account(request, account_id):
+    # 检查请求方法是否正确
     if request.method == 'DELETE':
         # 获取token
-        auth_token = request.META.get('HTTP_AUTHORIZATION').split(' ')[1]
+        temp = request.META.get('HTTP_AUTHORIZATION').split(' ')
+        if len(temp) > 1:
+            auth_token = request.META.get('HTTP_AUTHORIZATION').split(' ')[1]
+        else:
+            return JsonResponse({'message': 'Invalid AUTHORIZATION', 'status': 'error'})
+        # 寻找token对应的用户
         try:
             token = Token.objects.get(key=auth_token)
             current_user = token.user
         except Token.DoesNotExist:
             return JsonResponse({'message': 'Invalid AUTHORIZATION', 'status': 'error'})
+        # 检查当前用户是否为clerk
         if current_user.is_clerk():
             try:
                 user = User.objects.get(pk=account_id)
@@ -84,16 +107,20 @@ def delete_account(request, account_id):
                 return JsonResponse({'message': 'Account does not exist', 'status': 'error'})
 
         else:
-            return JsonResponse({'message': 'Current User is not clerk or admin', 'status': 'error'})
+            return JsonResponse({'message': 'Current User is not clerk', 'status': 'error'})
     else:
         return JsonResponse({'message': 'Invalid request method', 'status': 'error'})
 
 
-@csrf_exempt
 def get_balance(request):
     if request.method == 'GET':
         # 获取token
-        auth_token = request.META.get('HTTP_AUTHORIZATION').split(' ')[1]
+        temp = request.META.get('HTTP_AUTHORIZATION').split(' ')
+        if len(temp) > 1:
+            auth_token = request.META.get('HTTP_AUTHORIZATION').split(' ')[1]
+        else:
+            return JsonResponse({'message': 'Invalid AUTHORIZATION', 'status': 'error'})
+        # 寻找token对应的用户
         try:
             token = Token.objects.get(key=auth_token)
             user = token.user
@@ -104,8 +131,8 @@ def get_balance(request):
         return JsonResponse({'message': 'Invalid request method', 'status': 'error'})
 
 
-@csrf_exempt
 def transfer(request):
+    # 获取转账信息
     data = json.loads(request.body)
     destAccountId = data['destAccountId']
     amount = data['amount']
@@ -113,18 +140,25 @@ def transfer(request):
     if amount <= 0:
         return JsonResponse({'message': 'Invalid amount', 'status': 'error'})
     # 获取token
-    auth_token = request.META.get('HTTP_AUTHORIZATION').split(' ')[1]
+    temp = request.META.get('HTTP_AUTHORIZATION').split(' ')
+    if len(temp) > 1:
+        auth_token = request.META.get('HTTP_AUTHORIZATION').split(' ')[1]
+    else:
+        return JsonResponse({'message': 'Invalid AUTHORIZATION', 'status': 'error'})
+    # 寻找token对应的用户
     try:
         token = Token.objects.get(key=auth_token)
         current_user = token.user
     except Token.DoesNotExist:
         return JsonResponse({'message': 'Invalid AUTHORIZATION', 'status': 'error'})
+    # 检查余额是否足够
+    if current_user.balance < amount:
+        return JsonResponse({'message': 'Insufficient balance', 'status': 'error'})
+    # 查看目标账户是否存在
     try:
         dest_user = User.objects.get(pk=destAccountId)
     except User.DoesNotExist:
         return JsonResponse({'message': 'Destination account does not exist', 'status': 'error'})
-    if current_user.balance < amount:
-        return JsonResponse({'message': 'Insufficient balance', 'status': 'error'})
     current_user.balance -= amount
     dest_user.balance += amount
     current_user.save()
